Rnd generation of enemies complete and them advancing to player. Next step, cleanup enemies to be a new class for enemy groups (better logic), and then bullet shooting with collision with player

+++ rush00/Point.cpp
@@ -18,7 +18,7 @@ Point::Point() : x(0), y(0) {}
 
 /* :> Custom Constructor
  */
-Point::Point(int x, int y) : x(x), y(y) {}
+Point::Point(double x, double y) : x(x), y(y) {}
 
 /* :> Copy Constructor
 	- Assigns: the current class the values of the passed class.
